refactor: Rename `handler-decorator` to `instrument`
