--- node/packages/aws-lambda-sdk/.ts-types/handler-decorator.d.ts
@@ -1,2 +1,0 @@
-declare function handlerDecorator(value: Function): Function;
-export default handlerDecorator;
