+++ node/packages/aws-lambda-sdk/.ts-types/instrument.d.ts
@@ -1,0 +1,2 @@
+declare function instrument(value: Function): Function;
+export default instrument;
