+++ node/packages/aws-lambda-sdk/instrument.js
